Remover .pyc; add gitignore; Refatoramento de codigo
Teste git.

Refatoramento de código: limiar e limiar média ficam na questao6
